requirements.txt
django-environ==0.11.2
Django==5.0.0
requests==2.31.0
django-jsoneditor==0.2.4
djangorestframework==3.14.0
